% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_function.R
\name{boot_function}
\alias{boot_function}
\title{Bootstrapping for CIs Prevalence and Incidence}
\usage{
boot_function(
  data = sample_data,
  n_boot = 10000,
  weight_yn = F,
  weight_var = "weight",
  test_type_var = "type",
  record_id_var = "id",
  pi_init = 0.03,
  p_init = 0.01,
  se = 0.99,
  sp = 0.999,
  maxit = 1000,
  tol = 0.000001,
  seed = 1412414
)
}
\arguments{
\item{data}{input data}

\item{n_boot}{of bootstrapped samples}

\item{weight_yn}{whether used weighted data}

\item{weight_var}{variable for weights}

\item{test_type_var}{variable for scenario indicator}

\item{record_id_var}{patient id variable}

\item{pi_init}{prevalence to initialization}

\item{p_init}{incidence to initialization}

\item{se}{sensitivity/PPA}

\item{sp}{specificity/NPA}

\item{maxit}{maximum number iteriteration allowed}

\item{tol}{convergence criteria}

\item{seed}{starting seed}

\item{verbose}{show intermediate outputs}
}
\value{
Bootstrapped Confidence intervals.
}
\description{
This function calculates Confidence Intervals of MLE-EM estimates of Prevalence and Incidence
}
\examples{
boot_function(data=sample_data,n_boot=10000,weight_yn=F,weight_var="weight",test_type_var="type",record_id_var="id",pi_init=0.03,p_init=0.01,se=0.99, sp=0.999,maxit=1000,tol=1e-6,seed=1412414)
}
