% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_perform_subcohort.R
\name{combine_perform_subcohort}
\alias{combine_perform_subcohort}
\title{Combine The Performance on Population}
\usage{
combine_perform_subcohort(r, s, prob)
}
\arguments{
\item{r}{sensitivity/PPAs of combined Case definition in subpopulation}

\item{s}{specificity/NPA of combined Case definition in ith subpopulation}

\item{prob}{probability of each subpopulation in the entire cohort}
}
\value{
Calculate the Combined Performance of Case Definition in the entire populatioj.
}
\description{
This function calculate the S-matrix of the likelihood.
}
\examples{
combine_perform_subcohort(r=c(0.95,0.96,0.97,0.95,0.95,0.96),s=c(0.999,0.997,0.994,0.998,0.996,0.997),prob=c(0.1,0.2,0.3,0.2,0.1,0.1))
}
