% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_2parameters.R
\name{em_2parameters}
\alias{em_2parameters}
\title{MLE-EM estimates of Prevalence and Incidence}
\usage{
em_2parameters(
  nk = c(100, rep(200, 7), rep(150, 6)),
  pi_init = 0.03,
  p_init = 0.01,
  se = 0.99,
  sp = 0.999,
  maxit = 1000,
  tol = 1e-06,
  verbose = F
)
}
\arguments{
\item{nk}{input data}

\item{pi_init}{prevalence to initialization}

\item{p_init}{incidence to initialization}

\item{se}{sensitivity/PPA}

\item{sp}{specificity/NPA}

\item{maxit}{maximum number iteriteration allowed}

\item{tol}{convergence criteria}

\item{verbose}{show intermediate outputs}
}
\value{
MLE-EM estimates of Prevalence and Incidence.
}
\description{
This function calculates MLE-EM estimates of Prevalence and Incidence
}
\examples{
em_2parameters(nk=c(100, rep(200,7),rep(150,6)),pi_init=0.03,p_init=0.01,se=0.99, sp=0.999,maxit=1000,tol=1e-6,verbose=F)
}
