% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_function.R
\name{likelihood_function}
\alias{likelihood_function}
\title{S matrix}
\usage{
likelihood_function(se, sp, pi, p, nk)
}
\arguments{
\item{se}{sensitivity/PPA}

\item{sp}{sepecificity/NPA}

\item{pi}{prevalence}

\item{p}{incidence}

\item{nk}{input data}
}
\value{
Calculate the S-matrix of the likelihood.
}
\description{
This function calculate the S-matrix of the likelihood.
}
\examples{
likelihood_function(se=0.99,sp=0.9,pi=0.05,p=0.03,nk=c(6535,2816,2128,1929,1699,1399,1288, 41051,9273,8172,7227,6273,5430,4780))
}
