% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofu_function.R
\name{lofu_function}
\alias{lofu_function}
\title{Lost to Follow-up Function}
\usage{
lofu_function(n)
}
\arguments{
\item{n}{input data}
}
\value{
risk of lost to follow-up based on MLE approach.
}
\description{
This function calculates risk of LOFU.
}
\examples{
lofu_function(c(6535,2816,2128,1929,1699,1399,1288, 41051,9273,8172,7227,6273,5430,4780))
}
