% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_ci.R
\name{mle_ci}
\alias{mle_ci}
\title{MLE CIs Prevalence and Incidence}
\usage{
mle_ci(pi = 0.07, p = 0.02, nk, se, sp)
}
\arguments{
\item{pi}{adjusted prevalence}

\item{p}{adjusted incidence}

\item{se}{sensitivity/PPA}

\item{sp}{specificity/NPA}

\item{nK}{input data vector}
}
\value{
MLE confidence interval.
}
\description{
This function calculates Confidence Intervals of MLE-EM estimates of Prevalence and Incidence based on MLE
}
\examples{
mle_ci(pi=0.07,p=0.02,nk=,se=0.99,sp=0.999)
}
