% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesp_adj_prevalence.R
\name{sesp_adj_prevalence}
\alias{sesp_adj_prevalence}
\title{Combine The Performance From CZI Assays (1 PCR and 3 Ab Assays)}
\usage{
sesp_adj_prevalence(
  dat = dat_simulated,
  ab_nen_hyp_obs = 0.03,
  r1 = 0.99,
  s1 = 0.9999,
  r2 = 0.9,
  s2 = 0.999,
  r3 = 0.9,
  s3 = 0.999,
  r4 = 0.9,
  s4 = 0.999
)
}
\arguments{
\item{dat}{simulated baseline data on the patient level}

\item{ab_nen_hyp_obs}{Hypothesized Ab Neuralizing Observed Prevalence}

\item{r1}{sensitivity/PPA of PCR}

\item{s1}{specificity/NPA of PCR}

\item{r2}{sensitivity/PPA of Ab Spike (Stanford Ab Assay)}

\item{s2}{specificity/NPA of Ab Spike (Stanford Ab Assay)}

\item{r3}{sensitivity/PPA of Ab N-cap (UCSF Ab Assay)}

\item{s3}{specificity/NPA of Ab N-cap (UCSF Ab Assay)}

\item{r4}{sensitivity/PPA of Ab Neutralizing}

\item{s4}{specificity/NPA of Ab Neutralizing}
}
\value{
Calculate the Combined Performance of Case Definition.
}
\description{
This function calculate the S-matrix of the likelihood.
}
\examples{
sesp_adj_prevalence(dat=dat_simulated,ab_nen_hyp_obs=0.03,r1=0.99,s1=0.9999,r2=0.9,s2=0.999,r3=0.9,s3=0.999,r4=0.9,s4=0.999)
}
